+++ CHANGELOG.md
@@ -34,6 +34,7 @@ The following changes have been completed, and will be included in the next tagg
 - added `pickProperties()` helper function
 - added `getOwnKeys()` helper function
 - added `omitProperties()` helper function
+- added `HashMap.partial()` static helper method
 
 ### Refactor